Moved mock helper to test helpers
