+++ src/github/orbit-helpers.test.js
@@ -1,4 +1,5 @@
 import { refreshAuthTokens, isOAuthTokenExpired } from "../oauth_helpers";
+import { mockChromeStorage } from "../test-helpers";
 import {
   _getRepositoryFullName,
   _fetchRepositories,
@@ -7,63 +8,6 @@ import {
 } from "./orbit-helpers";
 
 jest.mock("../oauth_helpers");
-
-/**
- * Mocks the chrome storage object & getter
- * **Important**: At the end of your test, restore the original behaviour
- * by setting global.chrome to the object returned by this function.
- *
- * IE
- * const originalChrome = mockChromeStorage({ key: "123" })
- * ... Your tests
- * global.chrome = originalChrome
- *
- * @param {Object} objectToStore object to put in mock storage
- *
- * @returns {originalChrome} to restore behaviour to global.chrome
- */
-const mockChromeStorage = (objectToStore) => {
-  const mockChromeStorage = {
-    storage: objectToStore,
-    get: function (keys) {
-      // .get() should return all stored data
-      if (!keys) {
-        return this.storage;
-      }
-
-      const result = {};
-
-      // .get("repository_keys") should just return the stored data for "repository_keys"
-      if (typeof keys === "string") {
-        result[keys] = this.storage[keys];
-        return result;
-      }
-
-      // .get({ repository_keys: 123, repositories: 456) should fetch the data for repository_keys
-      // & repositories, and, if either is not found, return the default value given in the argument
-      Object.keys(keys).forEach((key) => {
-        result[key] = this.storage[key] || keys[key];
-      });
-
-      return result;
-    },
-    set: jest.fn((items) => {
-      Object.keys(items).forEach((key) => {
-        this.storage[key] = items[key];
-      });
-    }),
-  };
-
-  let originalChrome = global.chrome;
-
-  global.chrome = {
-    storage: {
-      sync: mockChromeStorage,
-    },
-  };
-
-  return originalChrome;
-};
 
 test("getOrbitCredentials should correctly configure orbit credentials object", async () => {
   const originalChrome = mockChromeStorage({
